fix: report config failures from the daemon and publish rsync stderr on every completed pass
Config load and reload failures now log at the daemon's own boundary with the
config path and the stage that failed, instead of four records inside the
loader; the reload arm also carries which trigger asked for the pass. A caller
that returned bare could exit silently on an empty /config.

Every terminal record that ends a completed transfer now publishes rsync's
captured stderr, not only the failure arms — the residual-error and vanished-
files paths computed the tail and discarded it, so an advisory on an exit-0
pass reached no reader.

Also in this change: the stat-then-open race is closed at all three config and
source read sites by checking the kind on the handle; remote uid/gid decode to
uint32 so a negative id is refused at the boundary rather than reaching rsync
as an unknown usermap name; --chown's published bound is corrected; and the
curl conversion in the rsync builder stage fails closed on an HTTP error.

app-review cycle 4: 44 plan items applied across 11 waves, 19 structural
reshapes, 14 findings voided. 19 challenges sustained cross-family, all
answered by amendment — four of them caught plan items that would have
implemented work the denial ledger had already refused.

+++ sync_args_test.go
@@ -10,7 +10,7 @@ import (
 
 // argJob returns a job with the given knobs set, using a fixed key path
 // (buildRsyncArgs does no filesystem access, so the path need not exist).
-func argJob(delete bool, uid, gid *int, excludes []string) *job {
+func argJob(delete bool, uid, gid *uint32, excludes []string) *job {
 	return &job{
 		Name:       "caddy",
 		Local:      "/sources/caddy",
@@ -60,7 +60,7 @@ func TestBuildRsyncArgs(t *testing.T) {
 
 	t.Run("uid only does not add chown", func(t *testing.T) {
 		t.Parallel()
-		got := buildRsyncArgs(argJob(false, new(1000), nil, nil), transport{})
+		got := buildRsyncArgs(argJob(false, new(uint32(1000)), nil, nil), transport{})
 		if hasChown(got) {
 			t.Errorf("--chown present with gid unset in %v", got)
 		}
@@ -68,7 +68,7 @@ func TestBuildRsyncArgs(t *testing.T) {
 
 	t.Run("gid only does not add chown", func(t *testing.T) {
 		t.Parallel()
-		got := buildRsyncArgs(argJob(false, nil, new(1000), nil), transport{})
+		got := buildRsyncArgs(argJob(false, nil, new(uint32(1000)), nil), transport{})
 		if hasChown(got) {
 			t.Errorf("--chown present with uid unset in %v", got)
 		}
@@ -112,7 +112,7 @@ func TestBuildRsyncArgs(t *testing.T) {
 
 	t.Run("all knobs together", func(t *testing.T) {
 		t.Parallel()
-		got := buildRsyncArgs(argJob(true, new(0), new(0), []string{"logs"}), transport{})
+		got := buildRsyncArgs(argJob(true, new(uint32(0)), new(uint32(0)), []string{"logs"}), transport{})
 		want := []string{
 			"-rlptD", "--delete", "--chown=0:0", "--stats", "-e", wantSSHAcceptNew,
 			"--filter=- .stfolder", "--filter=- .stversions",
@@ -390,7 +390,6 @@ func TestRemoteDest_bracketedIPv4Normalized(t *testing.T) {
 	}
 }
 
-
 // TestBuildRsyncArgs_transportSwitches pins the three opt-in transport
 // switches. The zero-value row is the important one: transport{} must append
 // none of -A/-X/-z, so no existing deployment changes behaviour on an image
@@ -463,4 +462,3 @@ func TestBuildRsyncArgs_transportSwitches(t *testing.T) {
 		})
 	}
 }
-
